requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.5.0; python_version >= '3.7'
click==8.0.3; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.73.0
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
packaging==21.3; python_version >= '3.6'
paho-mqtt==1.6.1
pydantic==1.9.0; python_full_version >= '3.6.1'
pyparsing==3.0.7; python_version >= '3.6'
redis==4.1.3
sniffio==1.2.0; python_version >= '3.5'
starlette==0.17.1; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
uvicorn==0.17.4
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
